秒杀商品程序锁，会出现超卖情况 原因：@Transactional事务会在开启它的方法结束后提交，如果在事务提交之前另一个方法进来读到了数据，那么读到的就是这个事务提交之前的数据

+++ seckill-common/src/main/java/com/seckill/common/bean/SeckillProduct.java
@@ -8,11 +8,11 @@ public class SeckillProduct {
 
     private Long productId;
 
-    private Integer seckillNum;
+    private Long seckillNum;
 
     private BigDecimal seckillPrice;
 
-    private Integer seckillInventory;
+    private Long seckillInventory;
 
     private Date createTime;
 
@@ -54,11 +54,11 @@ public class SeckillProduct {
         this.productId = productId;
     }
 
-    public Integer getSeckillNum() {
+    public Long getSeckillNum() {
         return seckillNum;
     }
 
-    public void setSeckillNum(Integer seckillNum) {
+    public void setSeckillNum(Long seckillNum) {
         this.seckillNum = seckillNum;
     }
 
@@ -70,11 +70,11 @@ public class SeckillProduct {
         this.seckillPrice = seckillPrice;
     }
 
-    public Integer getSeckillInventory() {
+    public Long getSeckillInventory() {
         return seckillInventory;
     }
 
-    public void setSeckillInventory(Integer seckillInventory) {
+    public void setSeckillInventory(Long seckillInventory) {
         this.seckillInventory = seckillInventory;
     }
